Fix emulation device
