+++ lib/connection.js
@@ -5,6 +5,7 @@ const { logEvent } = require('./logger');
 const { errorMessageForBrowserProcessCrash, closeBrowser } = require('./browserLauncher');
 const targetHandler = require('./handlers/targetHandler');
 const pageHandler = require('./handlers/pageHandler');
+const emulationHandler = require('./handlers/emulationHandler');
 const cri = require('chrome-remote-interface');
 const isReachable = require('is-reachable');
 const numRetries = defaultConfig.criConnectionRetries;
@@ -90,7 +91,7 @@ const initCRI = async (target, n, options = {}) => {
     }
     defaultConfig.device = process.env.TAIKO_EMULATE_DEVICE;
     if (defaultConfig.device) {
-      emulateDevice(defaultConfig.device);
+      emulationHandler.emulateDevice(defaultConfig.device);
     }
     await Promise.all(promises);
     eventHandler.removeAllListeners('handlerActingOnNewSession');
@@ -163,7 +164,7 @@ async function reconnect() {
 const cleanUpListenersOnClient = async () => {
   _client.removeAllListeners();
   await _client.close();
-}
+};
 
 const validate = () => {
   if (!_client) {
