+++ lib/handlers/emulationHandler.js
@@ -1,4 +1,5 @@
 const { eventHandler } = require('../eventBus');
+const networkHandler = require('./networkHandler');
 let emulation;
 
 const createdSessionListener = (client) => {
@@ -52,4 +53,17 @@ const setLocation = async (options) => {
   });
 };
 
-module.exports = { setLocation, setViewport, setTimeZone };
+const emulateDevice = async (deviceModel) => {
+  const devices = require('../data/devices').default;
+  const deviceEmulate = devices[deviceModel];
+  let deviceNames = Object.keys(devices);
+  if (deviceEmulate == undefined) {
+    throw new Error(`Please set one of the given device models \n${deviceNames.join('\n')}`);
+  }
+  await Promise.all([
+    setViewport(deviceEmulate.viewport),
+    networkHandler.setUserAgent(deviceEmulate),
+  ]);
+};
+
+module.exports = { setLocation, setViewport, setTimeZone, emulateDevice };
