+++ lib/taiko.js
@@ -320,16 +320,7 @@ module.exports.emulateNetwork = async (networkType) => {
 module.exports.emulateDevice = emulateDevice;
 async function emulateDevice(deviceModel) {
   validate();
-  const devices = require('./data/devices').default;
-  const deviceEmulate = devices[deviceModel];
-  let deviceNames = Object.keys(devices);
-  if (deviceEmulate == undefined) {
-    throw new Error(`Please set one of the given device models \n${deviceNames.join('\n')}`);
-  }
-  await Promise.all([
-    emulationHandler.setViewport(deviceEmulate.viewport),
-    networkHandler.setUserAgent(deviceEmulate),
-  ]);
+  await emulationHandler.emulateDevice(deviceModel);
   descEvent.emit('success', 'Device emulation set to ' + deviceModel);
 }
